Make a new class GeneralFileExtension
GeneralFileExtension is the class to be used by a script which produces a response, which is not uploaded to lims in a file. The response is exported to a file in test mode, and this file can also be used as the freezed reference. Let DriverFileExtension inherit from this class. Class methods are moved between these classes.