Preparing refactoring, split up Dilute class
Change name of Dilute class to SingleTransfer
Make class TransferEndpoint to be used instead of analyte as input to
SingleTransfer. In future when dealing with tube racks, analyte cannot
be used directly in SingleTransfer. In this stage, TransferEndpoint is
like an identical shell for analyte.

In addition, TransferEndpoint is needed as a unit after refactoring
RobotDeckPositioner