Add class SourceOnlyDilutionScheme
This is to be used in dilutions with only source analytes (qPCR). Here,
destination position doesn't matter. The driver file contains columns
for source positions and a fixed transfer volume. In addition, a small
fix in integration_test_service to import both input and output artifacts,
so that they can be updated in advance before the integration test starts.