Add concentration reference as input to DilutionScheme
Concentration reference is either ng/ul or nM (nano molar). Since calculations in DilutionScheme is not dependent on concentration unit, there is a variable 'concentration' that holds either of them. The field concentration is fetched from aliquot, either concentration_ngul or concentration_nm depending on conc ref.

DilutionScheme now have to be initialized from snpseq scripts with a concentration reference.