Fix fetching of SharedResultFile in pooling steps
1. Change initialization of artifact output_type
Apparently, a file intended to be a shared result file, has it's
attribute output-type changed from "SharedResultFile" to "ResultFile"
in cases when a step only has one sample attached to it. The change in
step_repository will resolve this.

2. (artifact_service.py)
In pool steps, analyte has generation_type = PER_ALL_INPUTS, so I
changed this.