Fix bug caused by shadowed name
The variable name `sample` was shadowing the parameter name `sample`.
Because of this, the loop changed the sample to some other sample and
finally returned the parent input artifact for that sample.

We saw this because the final sample was phix, but this could have
broken randomly for other samples too.