Validation results should only be shown once
Since different transfers will have the same validation results but in different
batches, sometimes there will be a duplication of the same validation message
in the logs. As a quick fix, add them to the set so they are only added once
per lifetime of the validation service. This works fine for our purposes since
we always use a fresh ValidationService for each extension. But it could be unexpected
if the service gets used in a different context, so it makes sense fixing this later.

Also add a minor fix in the end user notification.