Allow users to download a single shared file when there are two
Users can use a "local shared file" via the context, but this assumed there would be
only one file per file handle. We now often have more than one file. A temporary solution
is to allow the user to pass in "file_name_contains", which will differentiate between
the files based on the actual file name (not only the file handle in the Clarity UI).