Support zipping files to a file handle
When uploading files through the file service, one can now
zip them together. In that case, only one file needs to be available
on the file handle. If zip_files is not specified, the previous logic
will work as expected, where each of the files is uploaded to a single
file handle (which must be configured specially on the step.)

Closes #1601